[70197] Move node export processor nodes tree code to a separate class
