+++ Model/ImportExport/Processor/Export/Node.php
@@ -7,8 +7,7 @@ namespace Snowdog\Menu\Model\ImportExport\Processor\Export;
 use Magento\Framework\Api\SearchCriteriaBuilder;
 use Snowdog\Menu\Api\Data\NodeInterface;
 use Snowdog\Menu\Api\NodeRepositoryInterface;
-use Snowdog\Menu\Model\ImportExport\Processor\Export\Node\DataProcessor;
-use Snowdog\Menu\Model\ImportExport\Processor\ExtendedFields;
+use Snowdog\Menu\Model\ImportExport\Processor\Export\Node\Tree as NodeTree;
 
 class Node
 {
@@ -31,18 +30,18 @@ class Node
     private $nodeRepository;
 
     /**
-     * @var DataProcessor
+     * @var NodeTree
      */
-    private $dataProcessor;
+    private $nodeTree;
 
     public function __construct(
         SearchCriteriaBuilder $searchCriteriaBuilder,
         NodeRepositoryInterface $nodeRepository,
-        DataProcessor $dataProcessor
+        NodeTree $nodeTree
     ) {
         $this->searchCriteriaBuilder = $searchCriteriaBuilder;
         $this->nodeRepository = $nodeRepository;
-        $this->dataProcessor = $dataProcessor;
+        $this->nodeTree = $nodeTree;
     }
 
     public function getList(int $menuId): array
@@ -53,61 +52,6 @@ class Node
 
         $nodes = $this->nodeRepository->getList($searchCriteria)->getItems();
 
-        return $nodes ? $this->getNodesTree($nodes) : [];
-    }
-
-    private function getNodesTree(array $nodes): array
-    {
-        $tree = [];
-        $childNodesClusters = [];
-
-        foreach ($nodes as $node) {
-            $nodeId = $node->getId();
-            $parentId = $node->getParentId();
-            $nodeData = $this->dataProcessor->getData($node->getData());
-
-            $this->removeNodeExcludedFields($nodeData);
-
-            if (!$parentId) {
-                $tree[$nodeId] = $nodeData;
-                continue;
-            }
-
-            $childNodesClusters[$nodeId] = $nodeData;
-
-            if (isset($tree[$parentId])) {
-                $tree[$parentId][ExtendedFields::NODES][$nodeId] = &$childNodesClusters[$nodeId];
-                continue;
-            }
-
-            if (isset($childNodesClusters[$parentId])) {
-                $childNodesClusters[$parentId][ExtendedFields::NODES][$nodeId] = &$childNodesClusters[$nodeId];
-                continue;
-            }
-        }
-
-        return $this->reindexTreeNodes($tree);
-    }
-
-    private function removeNodeExcludedFields(array &$data): void
-    {
-        foreach (self::EXCLUDED_FIELDS as $excludedField) {
-            unset($data[$excludedField]);
-        }
-    }
-
-    private function reindexTreeNodes(array $nodes): array
-    {
-        $data = [];
-
-        foreach ($nodes as $node) {
-            if (isset($node[ExtendedFields::NODES])) {
-                $node[ExtendedFields::NODES] = $this->reindexTreeNodes($node[ExtendedFields::NODES]);
-            }
-
-            $data[] = $node;
-        }
-
-        return $data;
+        return $nodes ? $this->nodeTree->get($nodes) : [];
     }
 }
